Add example data for ingest and phylogenetic CI workflows [#12]
This is the full NCBI rubella dataset (taxon ID 11041) as downloaded
on 15 Apr 2025, and the ingested sequences and metadata as downloaded
from Nextstrain on 15 Apr 2025.

CI _could_ work just fine without embedding this data in the repo, but
having a hard-coded example-data file means that we don't depend on
NCBI or Nextstrain.org for successful CI runs.