WIP parse genotype from entrez
Downloads ENTREZ records, converts them to NDJSON and uses a custom
script to parse the genotype from the underlying GenBank record.

The download takes under 2min for c. 4k records. We were able to parse
genotypes from 3653/4181 records (87%), and we may be able to extract
genotypes from some of the remaining records if we spend more time
inspecting them.

See <https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1745366262183639>
for a discussion about the general approach of downloading ENTREZ and
converting to NDJSON. We may wish to take this further and switch the
curate pipeline to use ENTREZ-derived data, or we may wish to use ENTREZ
data for certain fields only (such as strain name, genotype etc).

Simple additions here are to parse the epiweek from the strain name and
use this as the sample date (range) when sample date is not known.

This commit is a WIP due to a few questions in the script and the usage
of the bio CLI to convert GenBank to NDJSON.